Create FileInfoLoader subclasses
and move to subdir
