+++ FileList/FileInfoLoader.cc
@@ -20,18 +20,41 @@
 #include <casacore/lattices/Lattices/HDF5Lattice.h>
 #include <casacore/mirlib/miriad.h>
 
-#include "../ImageData/Hdf5Attributes.h"
+#include "FitsInfoLoader.h"
+#include "GenericInfoLoader.h"
+#include "Hdf5InfoLoader.h"
+#include "PagedInfoLoader.h"
 
 //#################################################################################
 // FILE INFO LOADER
 
-FileInfoLoader::FileInfoLoader(const std::string& filename) : _filename(filename) {
-    _image_type = FileType(filename);
-}
-
-
-casacore::ImageOpener::ImageTypes FileInfoLoader::FileType(const std::string& file) {
-    return casacore::ImageOpener::imageType(file);
+FileInfoLoader* FileInfoLoader::GetInfoLoader(const std::string& filename) {
+    casacore::ImageOpener::ImageTypes image_type = casacore::ImageOpener::imageType(filename);
+    switch (image_type) {
+        case casacore::ImageOpener::AIPSPP:
+            return new PagedInfoLoader(filename, true);
+        case casacore::ImageOpener::FITS:
+            return new FitsInfoLoader(filename);
+        case casacore::ImageOpener::MIRIAD:
+            return new PagedInfoLoader(filename, false);
+        case casacore::ImageOpener::GIPSY:
+            break;
+        case casacore::ImageOpener::CAIPS:
+            break;
+        case casacore::ImageOpener::NEWSTAR:
+            break;
+        case casacore::ImageOpener::HDF5:
+            return new Hdf5InfoLoader(filename);
+        case casacore::ImageOpener::IMAGECONCAT:
+            break;
+        case casacore::ImageOpener::IMAGEEXPR:
+            break;
+        case casacore::ImageOpener::COMPLISTIMAGE:
+            break;
+        default:
+            break;
+    }
+    return new GenericInfoLoader(filename);
 }
 
 //#################################################################################
@@ -57,148 +80,16 @@ bool FileInfoLoader::FillFileInfo(CARTA::FileInfo* file_info) {
     }
 
     file_info->set_size(file_info_size);
-    file_info->set_type(ConvertFileType(_image_type));
+    CARTA::FileType file_type(GetCartaFileType());
+	file_info->set_type(file_type);
     casacore::String abs_file_name(cc_file.path().absoluteName());
     return GetHduList(file_info, abs_file_name);
 }
 
-CARTA::FileType FileInfoLoader::ConvertFileType(int cc_image_type) {
-    // convert casacore ImageType to protobuf FileType
-    switch (cc_image_type) {
-        case casacore::ImageOpener::FITS:
-            return CARTA::FileType::FITS;
-        case casacore::ImageOpener::AIPSPP:
-            return CARTA::FileType::CASA;
-        case casacore::ImageOpener::HDF5:
-            return CARTA::FileType::HDF5;
-        case casacore::ImageOpener::MIRIAD:
-            return CARTA::FileType::MIRIAD;
-        default:
-            return CARTA::FileType::UNKNOWN;
-    }
-}
-
 bool FileInfoLoader::GetHduList(CARTA::FileInfo* file_info, const std::string& filename) {
     // fill FileInfo hdu list
-    bool hdu_ok(false);
-    if (file_info->type() == CARTA::FileType::HDF5) {
-        casacore::HDF5File hdf_file(filename);
-        std::vector<casacore::String> hdus(casacore::HDF5Group::linkNames(hdf_file));
-        if (hdus.empty()) {
-            file_info->add_hdu_list("");
-            hdu_ok = true;
-        } else {
-            for (auto group_name : hdus)
-                file_info->add_hdu_list(group_name);
-            hdu_ok = true;
-        }
-    } else if (file_info->type() == CARTA::FITS) {
-        casacore::FitsInput fits_input(filename.c_str(), casacore::FITS::Disk, 10, FileInfoFitsErrHandler);
-        if (fits_input.err() == casacore::FitsIO::OK) { // check for cfitsio error
-            int num_hdu(fits_input.getnumhdu());
-            hdu_ok = (num_hdu > 0);
-            if (hdu_ok) {
-                // iterate through each header data unit
-                for (int hdu = 0; hdu < num_hdu; ++hdu) {
-                    if (fits_input.rectype() == casacore::FITS::HDURecord) {
-                        casacore::FITS::HDUType hdu_type = fits_input.hdutype();
-                        if (IsImageHdu(hdu_type)) {
-                            // add hdu to file info
-                            std::string hdu_name = std::to_string(hdu);
-                            int ndim(0);
-                            std::string ext_name;
-                            GetFitsHeaderInfo(fits_input, ndim, ext_name);
-                            if (ndim > 0) {
-                                if (!ext_name.empty()) {
-                                    hdu_name += " ExtName: " + ext_name;
-                                }
-                                file_info->add_hdu_list(hdu_name);
-                            }
-                            fits_input.skip_all(hdu_type); // skip data to next hdu
-                        } else {
-                            fits_input.skip_hdu();
-                        }
-                    }
-                }
-            }
-        }
-    } else {
-        file_info->add_hdu_list("");
-        hdu_ok = true;
-    }
-    return hdu_ok;
-}
-
-bool FileInfoLoader::IsImageHdu(casacore::FITS::HDUType hdu_type) {
-    return ((hdu_type == casacore::FITS::PrimaryArrayHDU) || (hdu_type == casacore::FITS::PrimaryGroupHDU) ||
-            (hdu_type == casacore::FITS::PrimaryTableHDU) || (hdu_type == casacore::FITS::ImageExtensionHDU));
-}
-
-void FileInfoLoader::GetFitsHeaderInfo(casacore::FitsInput& fits_input, int& ndim, std::string& ext_name) {
-    // return dims and extname from header
-    switch (fits_input.hdutype()) {
-        case casacore::FITS::PrimaryArrayHDU:
-        case casacore::FITS::PrimaryGroupHDU:
-        case casacore::FITS::PrimaryTableHDU: {
-            casacore::HeaderDataUnit* header_unit(nullptr);
-            switch (fits_input.datatype()) {
-                case casacore::FITS::CHAR:
-                    header_unit = new casacore::PrimaryArray<unsigned char>(fits_input);
-                    break;
-                case casacore::FITS::SHORT:
-                    header_unit = new casacore::PrimaryArray<short>(fits_input);
-                    break;
-                case casacore::FITS::LONG:
-                    header_unit = new casacore::PrimaryArray<int>(fits_input);
-                    break;
-                case casacore::FITS::FLOAT:
-                    header_unit = new casacore::PrimaryArray<float>(fits_input);
-                    break;
-                case casacore::FITS::DOUBLE:
-                    header_unit = new casacore::PrimaryArray<double>(fits_input);
-                    break;
-                default:
-                    break;
-            }
-            if (header_unit != nullptr) {
-                ndim = header_unit->dims();
-                delete header_unit;
-            }
-        } break;
-        case casacore::FITS::ImageExtensionHDU: {
-            switch (fits_input.datatype()) {
-                case casacore::FITS::CHAR: {
-                    casacore::ImageExtension<unsigned char> header_unit = casacore::ImageExtension<unsigned char>(fits_input);
-                    ndim = header_unit.dims();
-                    ext_name = header_unit.extname();
-                } break;
-                case casacore::FITS::SHORT: {
-                    casacore::ImageExtension<short> header_unit = casacore::ImageExtension<short>(fits_input);
-                    ndim = header_unit.dims();
-                    ext_name = header_unit.extname();
-                } break;
-                case casacore::FITS::LONG: {
-                    casacore::ImageExtension<int> header_unit = casacore::ImageExtension<int>(fits_input);
-                    ndim = header_unit.dims();
-                    ext_name = header_unit.extname();
-                } break;
-                case casacore::FITS::FLOAT: {
-                    casacore::ImageExtension<float> header_unit = casacore::ImageExtension<float>(fits_input);
-                    ndim = header_unit.dims();
-                    ext_name = header_unit.extname();
-                } break;
-                case casacore::FITS::DOUBLE: {
-                    casacore::ImageExtension<double> header_unit = casacore::ImageExtension<double>(fits_input);
-                    ndim = header_unit.dims();
-                    ext_name = header_unit.extname();
-                } break;
-                default:
-                    break;
-            }
-        } break;
-        default:
-            break;
-    }
+    file_info->add_hdu_list("");
+    return true;
 }
 
 //#################################################################################
@@ -210,694 +101,12 @@ bool FileInfoLoader::FillFileExtInfo(CARTA::FileInfoExtended* ext_info, std::str
         return false;
     }
 
-    bool ext_info_ok(false);
     std::string name(cc_file.path().baseName());
     auto entry = ext_info->add_computed_entries();
     entry->set_name("Name");
     entry->set_value(name);
     entry->set_entry_type(CARTA::EntryType::STRING);
-
-    // fill FileExtInfo depending on image type
-    switch (_image_type) {
-        case casacore::ImageOpener::AIPSPP:
-            ext_info_ok = FillCasaExtFileInfo(ext_info, message);
-            break;
-        case casacore::ImageOpener::FITS:
-            ext_info_ok = FillFitsExtFileInfo(ext_info, hdu, message);
-            break;
-        case casacore::ImageOpener::HDF5:
-            ext_info_ok = FillHdf5ExtFileInfo(ext_info, hdu, message);
-            break;
-        case casacore::ImageOpener::MIRIAD:
-            ext_info_ok = FillCasaExtFileInfo(ext_info, message);
-            break;
-        default:
-            break;
-    }
-    return ext_info_ok;
-}
-
-// HDF5
-bool FileInfoLoader::FillHdf5ExtFileInfo(CARTA::FileInfoExtended* ext_info, std::string& hdu, std::string& message) {
-    // Add extended info for HDF5 file
-    try {
-        auto hdf_file_ptr = casacore::CountedPtr<casacore::HDF5File>(new casacore::HDF5File(_filename));
-
-        if (hdu.empty()) { // use first
-            hdu = casacore::HDF5Group::linkNames(*hdf_file_ptr)[0];
-        }
-        casacore::HDF5Group hdf_group(*hdf_file_ptr, hdu, true);
-        casacore::Record attributes;
-        try {
-            // read attributes into casacore Record
-            attributes = Hdf5Attributes::ReadAttributes(hdf_group.getHid());
-        } catch (casacore::HDF5Error& err) {
-            message = "Error reading HDF5 attributes: " + err.getMesg();
-            return false;
-        }
-
-        if (attributes.empty()) {
-            message = "No HDF5 attributes";
-            return false;
-        }
-
-        // check dimensions
-        casacore::uInt num_dims;
-        casacore::IPosition data_shape;
-
-        try {
-            casacore::HDF5Lattice<float> hdf5_lattice = casacore::HDF5Lattice<float>(hdf_file_ptr, "DATA", hdu);
-            data_shape = hdf5_lattice.shape();
-            num_dims = data_shape.size();
-        } catch (casacore::AipsError& err) {
-            message = "Cannot open HDF5 DATA dataset.";
-            return false;
-        }
-        if (num_dims < 2 || num_dims > 4) {
-            message = "Image must be 2D, 3D or 4D.";
-            return false;
-        }
-
-        // extract values from Record
-        for (casacore::uInt field = 0; field < attributes.nfields(); ++field) {
-            auto header_entry = ext_info->add_header_entries();
-            header_entry->set_name(attributes.name(field));
-            switch (attributes.type(field)) {
-                case casacore::TpString: {
-                    *header_entry->mutable_value() = attributes.asString(field);
-                    header_entry->set_entry_type(CARTA::EntryType::STRING);
-                } break;
-                case casacore::TpBool: {
-                    casacore::Bool value_bool = attributes.asBool(field);
-                    *header_entry->mutable_value() = fmt::format("{}", value_bool);
-                    header_entry->set_entry_type(CARTA::EntryType::INT);
-                    header_entry->set_numeric_value(value_bool);
-                } break;
-                case casacore::TpInt:
-                case casacore::TpInt64: {
-                    casacore::Int64 value_int = attributes.asInt64(field);
-                    *header_entry->mutable_value() = fmt::format("{}", value_int);
-                    header_entry->set_entry_type(CARTA::EntryType::INT);
-                    header_entry->set_numeric_value(value_int);
-                } break;
-                case casacore::TpDouble: {
-                    casacore::Double numeric_val = attributes.asDouble(field);
-                    *header_entry->mutable_value() = fmt::format("{}", numeric_val);
-                    header_entry->set_entry_type(CARTA::EntryType::FLOAT);
-                    header_entry->set_numeric_value(numeric_val);
-                } break;
-                default:
-                    break;
-            }
-        }
-
-        // If in header, get values for computed entries:
-        // Get string values
-        std::string coord_type_x = (attributes.isDefined("CTYPE1") ? attributes.asString("CTYPE1") : "");
-        std::string coord_type_y = (attributes.isDefined("CTYPE2") ? attributes.asString("CTYPE2") : "");
-        std::string coord_type_3 = (attributes.isDefined("CTYPE3") ? attributes.asString("CTYPE3") : "");
-        std::string coord_type_4 = (attributes.isDefined("CTYPE4") ? attributes.asString("CTYPE4") : "");
-        std::string rade_sys = (attributes.isDefined("RADESYS") ? attributes.asString("RADESYS") : "");
-        std::string equinox = GetStringAttribute(attributes, "EQUINOX");
-        std::string spec_sys = (attributes.isDefined("SPECSYS") ? attributes.asString("SPECSYS") : "");
-        std::string bunit = (attributes.isDefined("BUNIT") ? attributes.asString("BUNIT") : "");
-        std::string crpix1 = GetStringAttribute(attributes, "CRPIX1");
-        std::string crpix2 = GetStringAttribute(attributes, "CRPIX2");
-        std::string cunit1 = (attributes.isDefined("CUNIT1") ? attributes.asString("CUNIT1") : "");
-        std::string cunit2 = (attributes.isDefined("CUNIT2") ? attributes.asString("CUNIT2") : "");
-        // Get numeric values
-        double crval1 = GetDoubleAttribute(attributes, "CRVAL1");
-        double crval2 = GetDoubleAttribute(attributes, "CRVAL2");
-        double cdelt1 = GetDoubleAttribute(attributes, "CDELT1");
-        double cdelt2 = GetDoubleAttribute(attributes, "CDELT2");
-        double bmaj = GetDoubleAttribute(attributes, "BMAJ");
-        double bmin = GetDoubleAttribute(attributes, "BMIN");
-        double bpa = GetDoubleAttribute(attributes, "BPA");
-
-        // shape, chan, stokes entries first
-        int chan_axis, stokes_axis;
-        FindChanStokesAxis(data_shape, coord_type_x, coord_type_y, coord_type_3, coord_type_4, chan_axis, stokes_axis);
-        AddShapeEntries(ext_info, data_shape, chan_axis, stokes_axis);
-        ext_info->add_stokes_vals(""); // not in header
-
-        // make computed entries strings
-        std::string xy_coords, cr_pixels, cr_coords, cr_deg_str, cr1, cr2, axis_inc, rs_beam;
-        if (!coord_type_x.empty() && !coord_type_y.empty())
-            xy_coords = fmt::format("{}, {}", coord_type_x, coord_type_y);
-        if (!crpix1.empty() && !crpix2.empty())
-            cr_pixels = fmt::format("[{}, {}] ", crpix1, crpix2);
-        if (!(crval1 == 0.0 && crval2 == 0.0))
-            cr_coords = fmt::format("[{:.4f} {}, {:.4f} {}]", crval1, cunit1, crval2, cunit2);
-        cr1 = MakeValueStr(coord_type_x, crval1, cunit1);
-        cr2 = MakeValueStr(coord_type_y, crval2, cunit2);
-        cr_deg_str = fmt::format("[{}, {}]", cr1, cr2);
-        if (!(cdelt1 == 0.0 && cdelt2 == 0.0))
-            axis_inc = fmt::format("{}, {}", UnitConversion(cdelt1, cunit1), UnitConversion(cdelt2, cunit2));
-        if (!(bmaj == 0.0 && bmin == 0.0 && bpa == 0.0))
-            rs_beam = Deg2Arcsec(bmaj) + " X " + Deg2Arcsec(bmin) + fmt::format(", {:.4f} deg", bpa);
-        MakeRadeSysStr(rade_sys, equinox);
-
-        // fill computed_entries
-        AddComputedEntries(ext_info, xy_coords, cr_pixels, cr_coords, cr_deg_str, rade_sys, spec_sys, bunit, axis_inc, rs_beam);
-    } catch (casacore::AipsError& err) {
-        message = "Error opening HDF5 file";
-        return false;
-    }
-    return true;
-}
-
-std::string FileInfoLoader::GetStringAttribute(casacore::Record& record, std::string field) {
-    // return attribute as string
-    std::string value;
-    if (record.isDefined(field)) {
-        switch (record.type(record.fieldNumber(field))) { // Hdf5Attributes only uses these types
-            case casacore::TpString: {
-                value = record.asString(field);
-            } break;
-            case casacore::TpInt64: {
-                value = std::to_string(record.asInt64(field));
-            } break;
-            case casacore::TpDouble: {
-                value = std::to_string(record.asDouble(field));
-            } break;
-            default:
-                break;
-        }
-    }
-    return value;
-}
-
-double FileInfoLoader::GetDoubleAttribute(casacore::Record& record, std::string field) {
-    // return attribute as double
-    double value(0.0);
-    if (record.isDefined(field)) {
-        switch (record.type(record.fieldNumber(field))) { // Hdf5Attributes only uses these types
-            case casacore::TpString: {
-                value = stod(record.asString(field));
-            } break;
-            case casacore::TpInt64: {
-                value = static_cast<double>(record.asInt64(field));
-            } break;
-            case casacore::TpDouble: {
-                value = record.asDouble(field);
-            } break;
-            default:
-                break;
-        }
-    }
-    return value;
-}
-
-// FITS
-bool FileInfoLoader::FillFitsExtFileInfo(CARTA::FileInfoExtended* ext_info, string& hdu, string& message) {
-    bool ext_info_ok(true);
-    try {
-        // convert string hdu to unsigned int
-        casacore::uInt hdu_num;
-        try {
-            casacore::String cc_hdu(hdu);
-            cc_hdu = cc_hdu.before(" "); // strip ExtName
-            cc_hdu.fromString(hdu_num, true);
-        } catch (casacore::AipsError& err) {
-            message = "Invalid hdu for FITS image.";
-            return false;
-        }
-
-        // check shape
-        casacore::Int num_dim(0);
-        casacore::IPosition data_shape;
-        try {
-            casacore::FITSImage fits_img(_filename, 0, hdu_num);
-            data_shape = fits_img.shape();
-            num_dim = data_shape.size();
-            if (num_dim < 2 || num_dim > 4) {
-                message = "Image must be 2D, 3D or 4D.";
-                return false;
-            }
-        } catch (casacore::AipsError& err) {
-            message = err.getMesg();
-            if (message.find("diagonal") != std::string::npos) { // "ArrayBase::diagonal() - diagonal out of range"
-                message = "Failed to open image at specified HDU.";
-            } else if (message.find("No image at specified location") != std::string::npos) {
-                message = "No image at specified HDU.";
-            } else {
-                message = "Failed to open image at specified HDU: " + message;
-            }
-            return false;
-        }
-        ext_info->set_dimensions(num_dim);
-
-        // use FITSTable to get Record of hdu entries
-        hdu_num += 1; // FITSTable starts at 1
-        casacore::FITSTable fits_table(_filename, hdu_num, true);
-        casacore::Record hdu_entries(fits_table.primaryKeywords().toRecord());
-        // set dims
-        ext_info->set_width(data_shape(0));
-        ext_info->set_height(data_shape(1));
-        ext_info->add_stokes_vals(""); // not in header
-
-        // set missing dims in header entries
-        if (!hdu_entries.isDefined("NAXIS")) {
-            auto header_entry = ext_info->add_header_entries();
-            header_entry->set_name("NAXIS");
-            *header_entry->mutable_value() = fmt::format("{}", num_dim);
-            header_entry->set_entry_type(CARTA::EntryType::INT);
-            header_entry->set_numeric_value(num_dim);
-        }
-        for (int i = 0; i < num_dim; ++i) {
-            casacore::String axis_name("NAXIS" + casacore::String::toString(i + 1));
-            if (!hdu_entries.isDefined(axis_name)) {
-                int naxis_i(data_shape(i));
-                auto header_entry = ext_info->add_header_entries();
-                header_entry->set_name(axis_name);
-                *header_entry->mutable_value() = fmt::format("{}", naxis_i);
-                header_entry->set_entry_type(CARTA::EntryType::INT);
-                header_entry->set_numeric_value(naxis_i);
-            }
-        }
-
-        // if in header, save values for computed entries
-        std::string coord_type_x, coord_type_y, coord_type3, coord_type4;
-        std::string rade_sys, equinox, spec_sys, bunit, crpix1, crpix2, cunit1, cunit2;
-        double crval1(0.0), crval2(0.0), cdelt1(0.0), cdelt2(0.0), bmaj(0.0), bmin(0.0), bpa(0.0);
-
-        // set header entries
-        for (casacore::uInt field = 0; field < hdu_entries.nfields(); ++field) {
-            casacore::String name = hdu_entries.name(field);
-            if ((name != "SIMPLE") && (name != "BITPIX") && !name.startsWith("PC")) {
-                auto header_entry = ext_info->add_header_entries();
-                header_entry->set_name(name);
-                casacore::DataType data_type(hdu_entries.type(field));
-                switch (data_type) {
-                    case casacore::TpString: {
-                        *header_entry->mutable_value() = hdu_entries.asString(field);
-                        header_entry->set_entry_type(CARTA::EntryType::STRING);
-                        if (header_entry->name() == "CTYPE1")
-                            coord_type_x = header_entry->value();
-                        else if (header_entry->name() == "CTYPE2")
-                            coord_type_y = header_entry->value();
-                        else if (header_entry->name() == "CTYPE3")
-                            coord_type3 = header_entry->value();
-                        else if (header_entry->name() == "CTYPE4")
-                            coord_type4 = header_entry->value();
-                        else if (header_entry->name() == "RADESYS")
-                            rade_sys = header_entry->value();
-                        else if (header_entry->name() == "SPECSYS")
-                            spec_sys = header_entry->value();
-                        else if (header_entry->name() == "BUNIT")
-                            bunit = header_entry->value();
-                        else if (header_entry->name() == "CUNIT1")
-                            cunit1 = header_entry->value();
-                        else if (header_entry->name() == "CUNIT2")
-                            cunit2 = header_entry->value();
-                        break;
-                    }
-                    case casacore::TpInt: {
-                        int64_t value_int(hdu_entries.asInt(field));
-                        if ((name == "NAXIS") && (value_int == 0))
-                            value_int = num_dim;
-                        *header_entry->mutable_value() = fmt::format("{}", value_int);
-                        header_entry->set_entry_type(CARTA::EntryType::INT);
-                        header_entry->set_numeric_value(value_int);
-                        break;
-                    }
-                    case casacore::TpFloat:
-                    case casacore::TpDouble: {
-                        double numeric_value(hdu_entries.asDouble(field));
-                        *header_entry->mutable_value() = fmt::format("{}", numeric_value);
-                        header_entry->set_entry_type(CARTA::EntryType::FLOAT);
-                        header_entry->set_numeric_value(numeric_value);
-                        if (header_entry->name() == "EQUINOX")
-                            equinox = std::to_string(static_cast<int>(numeric_value));
-                        else if (header_entry->name() == "CRVAL1")
-                            crval1 = numeric_value;
-                        else if (header_entry->name() == "CRVAL2")
-                            crval2 = numeric_value;
-                        else if (header_entry->name() == "CRPIX1")
-                            crpix1 = std::to_string(static_cast<int>(numeric_value));
-                        else if (header_entry->name() == "CRPIX2")
-                            crpix2 = std::to_string(static_cast<int>(numeric_value));
-                        else if (header_entry->name() == "CDELT1")
-                            cdelt1 = numeric_value;
-                        else if (header_entry->name() == "CDELT2")
-                            cdelt2 = numeric_value;
-                        else if (header_entry->name() == "BMAJ")
-                            bmaj = numeric_value;
-                        else if (header_entry->name() == "BMIN")
-                            bmin = numeric_value;
-                        else if (header_entry->name() == "BPA")
-                            bpa = numeric_value;
-                        break;
-                    }
-                    default:
-                        break;
-                }
-            }
-        }
-
-        // shape, chan, stokes entries first
-        int chan_axis, stokes_axis;
-        FindChanStokesAxis(data_shape, coord_type_x, coord_type_y, coord_type3, coord_type4, chan_axis, stokes_axis);
-        AddShapeEntries(ext_info, data_shape, chan_axis, stokes_axis);
-
-        // make strings for computed entries
-        std::string xy_coords, cr_pixels, cr_coords, cr1, cr2, cr_deg_str, axis_inc, rs_beam;
-        if (!coord_type_x.empty() && !coord_type_y.empty())
-            xy_coords = fmt::format("{}, {}", coord_type_x, coord_type_y);
-        if (!crpix1.empty() && !crpix2.empty())
-            cr_pixels = fmt::format("[{}, {}] ", crpix1, crpix2);
-        if (crval1 != 0.0 && crval2 != 0.0)
-            cr_coords = fmt::format("[{:.4f} {}, {:.4f} {}]", crval1, cunit1, crval2, cunit2);
-        cr1 = MakeValueStr(coord_type_x, crval1, cunit1);
-        cr2 = MakeValueStr(coord_type_y, crval2, cunit2);
-        cr_deg_str = fmt::format("[{}, {}]", cr1, cr2);
-        if (!(cdelt1 == 0.0 && cdelt2 == 0.0))
-            axis_inc = fmt::format("{}, {}", UnitConversion(cdelt1, cunit1), UnitConversion(cdelt2, cunit2));
-        if (!(bmaj == 0.0 && bmin == 0.0 && bpa == 0.0))
-            rs_beam = Deg2Arcsec(bmaj) + " X " + Deg2Arcsec(bmin) + fmt::format(", {:.4f} deg", bpa);
-        MakeRadeSysStr(rade_sys, equinox);
-
-        // fill computed_entries
-        AddComputedEntries(ext_info, xy_coords, cr_pixels, cr_coords, cr_deg_str, rade_sys, spec_sys, bunit, axis_inc, rs_beam);
-    } catch (casacore::AipsError& err) {
-        message = err.getMesg();
-        ext_info_ok = false;
-    }
-    return ext_info_ok;
-}
-
-// CASA, MIRIAD
-bool FileInfoLoader::FillCasaExtFileInfo(CARTA::FileInfoExtended* ext_info, std::string& message) {
-    bool ext_info_ok(true);
-    casacore::ImageInterface<casacore::Float>* cc_image(nullptr);
-    try {
-        switch (_image_type) {
-            case casacore::ImageOpener::AIPSPP: {
-                cc_image = new casacore::PagedImage<casacore::Float>(_filename);
-                break;
-            }
-            case casacore::ImageOpener::MIRIAD: {
-                // no way to catch error, use casacore miriad lib directly
-                int t_handle, i_handle, io_stat, num_dim;
-                hopen_c(&t_handle, _filename.c_str(), "old", &io_stat);
-                if (io_stat != 0) {
-                    message = "Could not open MIRIAD file";
-                    return false;
-                }
-                haccess_c(t_handle, &i_handle, "image", "read", &io_stat);
-                if (io_stat != 0) {
-                    message = "Could not open MIRIAD file";
-                    return false;
-                }
-                rdhdi_c(t_handle, "naxis", &num_dim, 0); // read "naxis" value into ndim, default 0
-                hdaccess_c(i_handle, &io_stat);
-                hclose_c(t_handle);
-                if (num_dim < 2 || num_dim > 4) {
-                    message = "Image must be 2D, 3D or 4D.";
-                    return false;
-                }
-                // hopefully okay to open now as casacore Image
-                cc_image = new casacore::MIRIADImage(_filename);
-                break;
-            }
-            default:
-                break;
-        }
-        if (cc_image == nullptr) {
-            message = "Unable to open image.";
-            return false;
-        }
-        // objects to retrieve header information
-        casacore::ImageInfo image_info(cc_image->imageInfo());
-        casacore::ImageSummary<casacore::Float> image_summary(*cc_image);
-        casacore::CoordinateSystem coord_sys(cc_image->coordinates());
-
-        // num_dims, shape
-        casacore::Int num_dims(image_summary.ndim());
-        ext_info->set_dimensions(num_dims);
-        if (num_dims < 2 || num_dims > 4) {
-            message = "Image must be 2D, 3D or 4D.";
-            return false;
-        }
-        casacore::IPosition data_shape(image_summary.shape());
-        ext_info->set_width(data_shape(0));
-        ext_info->set_height(data_shape(1));
-        ext_info->add_stokes_vals(""); // not in header
-        // set dims in header entries
-        auto num_axis_header_entry = ext_info->add_header_entries();
-        num_axis_header_entry->set_name("NAXIS");
-        *num_axis_header_entry->mutable_value() = fmt::format("{}", num_dims);
-        num_axis_header_entry->set_entry_type(CARTA::EntryType::INT);
-        num_axis_header_entry->set_numeric_value(num_dims);
-        for (casacore::Int i = 0; i < num_dims; ++i) {
-            auto header_entry = ext_info->add_header_entries();
-            header_entry->set_name("NAXIS" + casacore::String::toString(i + 1));
-            *header_entry->mutable_value() = fmt::format("{}", data_shape(i));
-            header_entry->set_entry_type(CARTA::EntryType::INT);
-            header_entry->set_numeric_value(data_shape(i));
-        }
-
-        // if in header, save values for computed entries
-        std::string rs_beam, bunit, projection, equinox, rade_sys, coord_type_x, coord_type_y, coord_type3, coord_type4, spec_sys;
-
-        // BMAJ, BMIN, BPA
-        if (image_info.hasBeam() && image_info.hasSingleBeam()) {
-            // get values
-            casacore::GaussianBeam rbeam(image_info.restoringBeam());
-            casacore::Quantity major_axis(rbeam.getMajor()), minor_axis(rbeam.getMinor()), pa(rbeam.getPA(true));
-            major_axis.convert("deg");
-            minor_axis.convert("deg");
-            pa.convert("deg");
-
-            // add to header entries
-            casacore::Double bmaj(major_axis.getValue()), bmin(minor_axis.getValue());
-            casacore::Float bpa(pa.getValue());
-            num_axis_header_entry = ext_info->add_header_entries();
-            num_axis_header_entry->set_name("BMAJ");
-            *num_axis_header_entry->mutable_value() = fmt::format("{}", bmaj);
-            num_axis_header_entry->set_entry_type(CARTA::EntryType::FLOAT);
-            num_axis_header_entry->set_numeric_value(bmaj);
-            num_axis_header_entry = ext_info->add_header_entries();
-            num_axis_header_entry->set_name("BMIN");
-            *num_axis_header_entry->mutable_value() = fmt::format("{}", bmin);
-            num_axis_header_entry->set_entry_type(CARTA::EntryType::FLOAT);
-            num_axis_header_entry->set_numeric_value(bmin);
-            num_axis_header_entry = ext_info->add_header_entries();
-            num_axis_header_entry->set_name("BPA");
-            *num_axis_header_entry->mutable_value() = fmt::format("{}", bpa);
-            num_axis_header_entry->set_entry_type(CARTA::EntryType::FLOAT);
-            num_axis_header_entry->set_numeric_value(bpa);
-
-            // add to computed entries
-            rs_beam = fmt::format("{} X {}, {:.4f} deg", Deg2Arcsec(bmaj), Deg2Arcsec(bmin), bpa);
-        }
-
-        // type
-        num_axis_header_entry = ext_info->add_header_entries();
-        num_axis_header_entry->set_name("BTYPE");
-        *num_axis_header_entry->mutable_value() = casacore::ImageInfo::imageType(image_info.imageType());
-        num_axis_header_entry->set_entry_type(CARTA::EntryType::STRING);
-        // object
-        num_axis_header_entry = ext_info->add_header_entries();
-        num_axis_header_entry->set_name("OBJECT");
-        *num_axis_header_entry->mutable_value() = image_info.objectName();
-        num_axis_header_entry->set_entry_type(CARTA::EntryType::STRING);
-        // units
-        num_axis_header_entry = ext_info->add_header_entries();
-        num_axis_header_entry->set_name("BUNIT");
-        bunit = image_summary.units().getName();
-        *num_axis_header_entry->mutable_value() = bunit;
-        num_axis_header_entry->set_entry_type(CARTA::EntryType::STRING);
-
-        // Direction axes: projection, equinox, radesys
-        casacore::Vector<casacore::String> dir_axis_names; // axis names to append projection
-        casacore::MDirection::Types dir_type;
-        casacore::Int i_dir_coord(coord_sys.findCoordinate(casacore::Coordinate::DIRECTION));
-        if (i_dir_coord >= 0) {
-            const casacore::DirectionCoordinate& dir_coord = coord_sys.directionCoordinate(casacore::uInt(i_dir_coord));
-            // direction axes
-            projection = dir_coord.projection().name();
-            dir_axis_names = dir_coord.worldAxisNames();
-            // equinox, radesys
-            dir_coord.getReferenceConversion(dir_type);
-            equinox = casacore::MDirection::showType(dir_type);
-            GetRadeSysFromEquinox(rade_sys, equinox);
-        }
-
-        // get imSummary axes values: name, reference pixel and value, pixel increment, units
-        casacore::Vector<casacore::String> ax_names(image_summary.axisNames());
-        casacore::Vector<casacore::Double> ax_ref_pix(image_summary.referencePixels());
-        casacore::Vector<casacore::Double> ax_ref_val(image_summary.referenceValues());
-        casacore::Vector<casacore::Double> ax_increments(image_summary.axisIncrements());
-        casacore::Vector<casacore::String> ax_units(image_summary.axisUnits());
-        size_t axis_size(ax_names.size());
-        for (casacore::uInt i = 0; i < axis_size; ++i) {
-            casacore::String suffix(casacore::String::toString(i + 1)); // append to keyword name
-            casacore::String axis_name = ax_names(i);
-            // modify direction axes, if any
-            if ((i_dir_coord >= 0) && anyEQ(dir_axis_names, axis_name)) {
-                ConvertAxisName(axis_name, projection, dir_type); // FITS name with projection
-                if (ax_units(i) == "rad") {                       // convert ref value and increment to deg
-                    casacore::Quantity ref_val_quant(ax_ref_val(i), ax_units(i));
-                    ref_val_quant.convert("deg");
-                    casacore::Quantity increment_quant(ax_increments(i), ax_units(i));
-                    increment_quant.convert("deg");
-                    // update values to use later for computed entries
-                    ax_ref_val(i) = ref_val_quant.getValue();
-                    ax_increments(i) = increment_quant.getValue();
-                    ax_units(i) = increment_quant.getUnit();
-                }
-            }
-            // name = CTYPE
-            // add header entry
-            num_axis_header_entry = ext_info->add_header_entries();
-            num_axis_header_entry->set_name("CTYPE" + suffix);
-            num_axis_header_entry->set_entry_type(CARTA::EntryType::STRING);
-            *num_axis_header_entry->mutable_value() = axis_name;
-            // computed entries
-            if (suffix == "1")
-                coord_type_x = axis_name;
-            else if (suffix == "2")
-                coord_type_y = axis_name;
-            else if (suffix == "3")
-                coord_type3 = axis_name;
-            else if (suffix == "4")
-                coord_type4 = axis_name;
-
-            // ref val = CRVAL
-            num_axis_header_entry = ext_info->add_header_entries();
-            num_axis_header_entry->set_name("CRVAL" + suffix);
-            *num_axis_header_entry->mutable_value() = fmt::format("{}", ax_ref_val(i));
-            num_axis_header_entry->set_entry_type(CARTA::EntryType::FLOAT);
-            num_axis_header_entry->set_numeric_value(ax_ref_val(i));
-            // increment = CDELT
-            num_axis_header_entry = ext_info->add_header_entries();
-            num_axis_header_entry->set_name("CDELT" + suffix);
-            *num_axis_header_entry->mutable_value() = fmt::format("{}", ax_increments(i));
-            num_axis_header_entry->set_entry_type(CARTA::EntryType::FLOAT);
-            num_axis_header_entry->set_numeric_value(ax_increments(i));
-            // ref pix = CRPIX
-            num_axis_header_entry = ext_info->add_header_entries();
-            num_axis_header_entry->set_name("CRPIX" + suffix);
-            *num_axis_header_entry->mutable_value() = fmt::format("{}", ax_ref_pix(i));
-            num_axis_header_entry->set_entry_type(CARTA::EntryType::FLOAT);
-            num_axis_header_entry->set_numeric_value(ax_ref_pix(i));
-            // units = CUNIT
-            num_axis_header_entry = ext_info->add_header_entries();
-            num_axis_header_entry->set_name("CUNIT" + suffix);
-            *num_axis_header_entry->mutable_value() = ax_units(i);
-            num_axis_header_entry->set_entry_type(CARTA::EntryType::STRING);
-        }
-
-        // RESTFRQ
-        casacore::String rest_freq_string;
-        casacore::Quantum<casacore::Double> rest_freq;
-        casacore::Bool ok = image_summary.restFrequency(rest_freq_string, rest_freq);
-        if (ok) {
-            num_axis_header_entry = ext_info->add_header_entries();
-            num_axis_header_entry->set_name("RESTFRQ");
-            casacore::Double rest_freq_val(rest_freq.getValue());
-            *num_axis_header_entry->mutable_value() = rest_freq_string;
-            num_axis_header_entry->set_entry_type(CARTA::EntryType::FLOAT);
-            num_axis_header_entry->set_numeric_value(rest_freq_val);
-        }
-
-        // SPECSYS
-        casacore::Int i_spec_coord(coord_sys.findCoordinate(casacore::Coordinate::SPECTRAL));
-        if (i_spec_coord >= 0) {
-            const casacore::SpectralCoordinate& spectral_coordinate = coord_sys.spectralCoordinate(casacore::uInt(i_spec_coord));
-            casacore::MFrequency::Types freq_type;
-            casacore::MEpoch epoch;
-            casacore::MPosition position;
-            casacore::MDirection direction;
-            spectral_coordinate.getReferenceConversion(freq_type, epoch, position, direction);
-            bool have_spec_sys = ConvertSpecSysToFits(spec_sys, freq_type);
-            if (have_spec_sys) {
-                num_axis_header_entry = ext_info->add_header_entries();
-                num_axis_header_entry->set_name("SPECSYS");
-                *num_axis_header_entry->mutable_value() = spec_sys;
-                num_axis_header_entry->set_entry_type(CARTA::EntryType::STRING);
-            }
-        }
-
-        // RADESYS, EQUINOX
-        if (!rade_sys.empty()) {
-            num_axis_header_entry = ext_info->add_header_entries();
-            num_axis_header_entry->set_name("RADESYS");
-            *num_axis_header_entry->mutable_value() = rade_sys;
-            num_axis_header_entry->set_entry_type(CARTA::EntryType::STRING);
-        }
-        if (!equinox.empty()) {
-            num_axis_header_entry = ext_info->add_header_entries();
-            num_axis_header_entry->set_name("EQUINOX");
-            *num_axis_header_entry->mutable_value() = equinox;
-            num_axis_header_entry->set_entry_type(CARTA::EntryType::STRING);
-        }
-        // computed entries
-        MakeRadeSysStr(rade_sys, equinox);
-
-        // Other summary items
-        // telescope
-        num_axis_header_entry = ext_info->add_header_entries();
-        num_axis_header_entry->set_name("TELESCOP");
-        *num_axis_header_entry->mutable_value() = image_summary.telescope();
-        num_axis_header_entry->set_entry_type(CARTA::EntryType::STRING);
-        // observer
-        num_axis_header_entry = ext_info->add_header_entries();
-        num_axis_header_entry->set_name("OBSERVER");
-        *num_axis_header_entry->mutable_value() = image_summary.observer();
-        num_axis_header_entry->set_entry_type(CARTA::EntryType::STRING);
-        // obs date
-        casacore::MEpoch epoch;
-        num_axis_header_entry = ext_info->add_header_entries();
-        num_axis_header_entry->set_name("DATE");
-        *num_axis_header_entry->mutable_value() = image_summary.obsDate(epoch);
-        num_axis_header_entry->set_entry_type(CARTA::EntryType::STRING);
-
-        // shape, chan, stokes entries first
-        int chan_axis, stokes_axis;
-        FindChanStokesAxis(data_shape, coord_type_x, coord_type_y, coord_type3, coord_type4, chan_axis, stokes_axis);
-        AddShapeEntries(ext_info, data_shape, chan_axis, stokes_axis);
-
-        // computed_entries
-        std::string xy_coords, cr_pixels, cr_coords, cr_deg_str, axis_inc;
-        int cr_pix0, cr_pix1;
-        std::string cunit0, cr0, cunit1, cr1;
-        double crval0, crval1, cdelt0, cdelt1;
-        if (axis_size >= 1) { // pv images only have first axis values
-            cr_pix0 = static_cast<int>(ax_ref_pix(0));
-            cunit0 = ax_units(0);
-            crval0 = ax_ref_val(0);
-            cdelt0 = ax_increments(0);
-            cr0 = MakeValueStr(coord_type_x, crval0, cunit0); // for angle format
-            if (axis_size > 1) {
-                cr_pix1 = static_cast<int>(ax_ref_pix(1));
-                cunit1 = ax_units(1);
-                crval1 = ax_ref_val(1);
-                cdelt1 = ax_increments(1);
-                cr1 = MakeValueStr(coord_type_y, crval1, cunit1); // for angle format
-                xy_coords = fmt::format("{}, {}", coord_type_x, coord_type_y);
-                cr_pixels = fmt::format("[{}, {}]", cr_pix0, cr_pix1);
-                cr_coords = fmt::format("[{:.3f} {}, {:.3f} {}]", crval0, cunit0, crval1, cunit1);
-                cr_deg_str = fmt::format("[{} {}]", cr0, cr1);
-                axis_inc = fmt::format("{}, {}", UnitConversion(cdelt0, cunit0), UnitConversion(cdelt1, cunit1));
-            } else {
-                xy_coords = fmt::format("{}", coord_type_x);
-                cr_pixels = fmt::format("[{}]", cr_pix0);
-                cr_coords = fmt::format("[{:.3f} {}]", crval0, cunit0);
-                cr_deg_str = fmt::format("[{}]", cr0);
-                axis_inc = fmt::format("{}", UnitConversion(cdelt0, cunit0));
-            }
-        }
-        AddComputedEntries(ext_info, xy_coords, cr_pixels, cr_coords, cr_deg_str, rade_sys, spec_sys, bunit, axis_inc, rs_beam);
-    } catch (casacore::AipsError& err) {
-        delete cc_image;
-        message = err.getMesg().c_str();
-        ext_info_ok = false;
-    }
-
-    delete cc_image;
-    return ext_info_ok;
+    return FillExtFileInfo(ext_info, hdu, message);
 }
 
 // ***** Computed entries *****
@@ -1091,47 +300,6 @@ void FileInfoLoader::FindChanStokesAxis(const casacore::IPosition& data_shape, c
 
 // ***** FITS keyword conversion *****
 
-void FileInfoLoader::ConvertAxisName(std::string& axis_name, std::string& projection, casacore::MDirection::Types type) {
-    // Convert direction axis name to FITS name and append projection
-    if ((axis_name == "Right Ascension") || (axis_name == "Hour Angle")) {
-        axis_name = (projection.empty() ? "RA" : "RA---" + projection);
-    } else if (axis_name == "Declination") {
-        axis_name = (projection.empty() ? "DEC" : "DEC--" + projection);
-    } else if (axis_name == "Longitude") {
-        switch (type) {
-            case casacore::MDirection::GALACTIC:
-                axis_name = (projection.empty() ? "GLON" : "GLON-" + projection);
-                break;
-            case casacore::MDirection::SUPERGAL:
-                axis_name = (projection.empty() ? "SLON" : "SLON-" + projection);
-                break;
-            case casacore::MDirection::ECLIPTIC:
-            case casacore::MDirection::MECLIPTIC:
-            case casacore::MDirection::TECLIPTIC:
-                axis_name = (projection.empty() ? "ELON" : "ELON-" + projection);
-                break;
-            default:
-                break;
-        }
-    } else if (axis_name == "Latitude") {
-        switch (type) {
-            case casacore::MDirection::GALACTIC:
-                axis_name = (projection.empty() ? "GLAT" : "GLAT-" + projection);
-                break;
-            case casacore::MDirection::SUPERGAL:
-                axis_name = (projection.empty() ? "SLAT" : "SLAT-" + projection);
-                break;
-            case casacore::MDirection::ECLIPTIC:
-            case casacore::MDirection::MECLIPTIC:
-            case casacore::MDirection::TECLIPTIC:
-                axis_name = (projection.empty() ? "ELAT" : "ELAT-" + projection);
-                break;
-            default:
-                break;
-        }
-    }
-}
-
 void FileInfoLoader::MakeRadeSysStr(std::string& rade_sys, const std::string& equinox) {
     // append equinox to radesys
     std::string prefix;
@@ -1144,18 +312,6 @@ void FileInfoLoader::MakeRadeSysStr(std::string& rade_sys, const std::string& eq
     if (!rade_sys.empty() && !equinox.empty())
         rade_sys.append(", ");
     rade_sys.append(prefix + equinox);
-}
-
-void FileInfoLoader::GetRadeSysFromEquinox(std::string& rade_sys, std::string& equinox) {
-    // according to casacore::FITSCoordinateUtil::toFITSHeader
-    if (equinox.find("ICRS") != std::string::npos) {
-        rade_sys = "ICRS";
-        equinox = "2000";
-    } else if (equinox.find("2000") != std::string::npos) {
-        rade_sys = "FK5";
-    } else if (equinox.find("B1950") != std::string::npos) {
-        rade_sys = "FK4";
-    }
 }
 
 std::string FileInfoLoader::MakeValueStr(const std::string& type, double val, const std::string& unit) {
@@ -1183,44 +339,6 @@ std::string FileInfoLoader::MakeValueStr(const std::string& type, double val, co
         }
     }
     return val_str;
-}
-
-bool FileInfoLoader::ConvertSpecSysToFits(std::string& spec_sys, casacore::MFrequency::Types type) {
-    // use labels in FITS headers
-    bool result(true);
-    switch (type) {
-        case casacore::MFrequency::LSRK:
-            spec_sys = "LSRK";
-            break;
-        case casacore::MFrequency::BARY:
-            spec_sys = "BARYCENT";
-            break;
-        case casacore::MFrequency::LSRD:
-            spec_sys = "LSRD";
-            break;
-        case casacore::MFrequency::GEO:
-            spec_sys = "GEOCENTR";
-            break;
-        case casacore::MFrequency::REST:
-            spec_sys = "SOURCE";
-            break;
-        case casacore::MFrequency::GALACTO:
-            spec_sys = "GALACTOC";
-            break;
-        case casacore::MFrequency::LGROUP:
-            spec_sys = "LOCALGRP";
-            break;
-        case casacore::MFrequency::CMB:
-            spec_sys = "CMBDIPOL";
-            break;
-        case casacore::MFrequency::TOPO:
-            spec_sys = "TOPOCENT";
-            break;
-        default:
-            spec_sys = "";
-            result = false;
-    }
-    return result;
 }
 
 // ***** Unit conversion *****
