+++ Session.h
@@ -203,8 +203,8 @@ private:
     FileListHandler* _file_list_handler;
 
     // File info for browser, open file
-    CARTA::FileInfo* _selected_file_info;
-    CARTA::FileInfoExtended* _selected_file_info_extended;
+	std::unique_ptr<CARTA::FileInfo> _selected_file_info;
+	std::unique_ptr<CARTA::FileInfoExtended> _selected_file_info_extended;
 
     // Frame
     std::unordered_map<int, std::unique_ptr<Frame>> _frames; // <file_id, Frame>: one frame per image file
